dct: force float promotion in the IDCT dequantization (sm_1x fix)
The IDCT kernel multiplies int16 coefficients by the uint16
quantization table.  Under nvcc's legacy sm_1x frontend
(Open64/nvopencc - everything below compute 2.0) the int16 operand
is loaded UNSIGNED for this mixed multiply: negative coefficients
read as 65536+x and the reconstruction saturates into ~5 dB noise.
The NVVM frontend (sm_20+) gets it right, which is why the bug only
ever shows on cc 1.x devices.

Cast both operands to float explicitly, forcing cvt.f32.s16 on the
coefficient load.  Verified bit-exact against the host-reference
IDCT via a staged kernel harness (the dequant-stage dump showed the
exact uint16 reinterpretations; ptxas -O0 exonerated optimization),
and end-to-end GPU decode PSNR matches ffmpeg's decode of the same
JPEG to 0.1 dB on a GeForce GT 130 (compute 1.1, CUDA 6.5).

This is the bug behind the 2014 "currently broken on cc < 2.0"
gate: the gate landed right after the 2013 IDCT rewrite that
introduced the mixed-width multiply.

+++ src/gpujpeg_dct_gpu.cu
@@ -495,9 +495,15 @@ gpujpeg_idct_gpu_kernel(int16_t* source, uint8_t* result, int output_stride, uin
 	//so this proves to be the fastest way
 #pragma unroll
 	for (int i = 0; i < 8; i++) {
-		data[threadIdx.z][i][threadIdx.x][threadIdx.y] = sourcePtr[x8
+		// Explicit float promotion of BOTH operands: the int16_t x uint16_t
+		// integer multiply tempts nvcc's legacy sm_1x frontend (Open64) into
+		// an unsigned 16-bit load of the coefficient - negative coefficients
+		// become huge positives and the IDCT saturates (GT 130 / CUDA 6.5,
+		// found 2026-07-20).  Verified bit-exact on sm_11 with the casts;
+		// modern NVVM codegen is unaffected either way.
+		data[threadIdx.z][i][threadIdx.x][threadIdx.y] = (float) sourcePtr[x8
 				+ threadIdx.y + i * GPUJPEG_IDCT_BLOCK_X * GPUJPEG_IDCT_BLOCK_Y + z64 * 8]
-				* quantization_table[threadIdx.x * 8 + threadIdx.y];
+				* (float) quantization_table[threadIdx.x * 8 + threadIdx.y];
 	}
 	
 	__syncthreads();
